+++ migrations/versions/d94f80b3edcd_add_phone_to_accounts.py
@@ -7,6 +7,7 @@ Create Date: 2026-03-20 03:06:56.517034
 """
 from alembic import op
 import sqlalchemy as sa
+from sqlalchemy import inspect
 
 
 # revision identifiers, used by Alembic.
@@ -16,11 +17,17 @@ branch_labels = None
 depends_on = None
 
 
+def _column_exists(table, column):
+    return column in [c['name'] for c in inspect(op.get_bind()).get_columns(table)]
+
+
 def upgrade():
-    with op.batch_alter_table('accounts', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('phone', sa.String(length=20), nullable=True))
+    if not _column_exists('accounts', 'phone'):
+        with op.batch_alter_table('accounts', schema=None) as batch_op:
+            batch_op.add_column(sa.Column('phone', sa.String(length=20), nullable=True))
 
 
 def downgrade():
-    with op.batch_alter_table('accounts', schema=None) as batch_op:
-        batch_op.drop_column('phone')
+    if _column_exists('accounts', 'phone'):
+        with op.batch_alter_table('accounts', schema=None) as batch_op:
+            batch_op.drop_column('phone')
